backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.6.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
